Refactored widgets to new files
